Added npm default config.
This patch adds an npmrc configuration file, which clears the
default version tagging prefix. This brings this project in line
with the standard tagging method used in openstack, rather than
'v1.0.0', it will use '1.0.0' as the tag. It also forces npm to
attempt to gpg-sign any tags created via `npm version`.

Change-Id: If7679f579b52633911bc5d6ee1f46dfae0104754