Enable brace-style
Require Brace Style (brace-style)

Brace style is closely related to indent style in programming and
describes the placement of curly braces relative to their control
statement and body. There are probably a dozen, if not more, brace
styles in the world.

The one true brace style is one of the most common brace styles in
JavaScript, in which the opening curly brace of a block is placed on
the same line as its corresponding statement or declaration.

http://eslint.org/docs/rules/brace-style

Change-Id: I2902ef3821f7a2e89f24b8e8af542919c54c24fe