Enable strict
This enables the function form of the "use strict" invocation. In
general, 'strict' is considered a good practice, so enabling it
is good. Requiring the function method of use strict avoid a
common bug encountered with a global strict declaration, where
the rule applies to all javascript that comes after the global
"use strict" statement. Using the function method prevents
brittleness, by creating a scope around the 'strict' case.

http://eslint.org/docs/rules/strict

Change-Id: I5c89c26d2d7a38c342e7c8f49f8cabe145dd02c0