Clarified wording in governance
This patch clarifies the wording on a few governance statements.
First, the intent of the rule approval policy is to encourage more
pairs of eyeballs - thus the wording was changed from +5 to 'five
positive votes'. Secondly - library upgrades should _never_ change
the existing rule set, because they already impose changes due
to bugfixes in the rules themselves.

Change-Id: I4f2765bcd04cb50da94ac121259c0cc8244bda19