Fix building of the pdf documentation
The .pdf documentation could not be build without errors because the
source code of the documentation contained errors. In addition, there
were other problems:

    - The readme section of the documentation was in .md format
      instead of .rst format. As a result, the created documentation did
      not look good.
    - The .html documentation used deprecated oslosphinx theme instead
      of openstackdocstheme.

This patch fixes the above-mentioned problems and makes sure that
the documentation is generated properly.

Also, this patch updates nodejs4-docs to nodejs10-docs because the
nodejs4-docs was using deprecated nodejs-npm-run-docs job [1].

[1] https://review.opendev.org/c/openstack/openstack-zuul-jobs/+/757101

Task: 35462
Depends-On: I738b833109e4caeb58bb391d79d6e63284462bd8
Change-Id: I921b15edda433c3e47456488da6d2bda07c34262